requirements.txt
ag2==0.9.7
aiofiles==24.1.0
annotated-types==0.7.0
anyio==4.10.0
asyncer==0.0.8
autogen==0.9.7
autogen-agentchat==0.7.1
autogen-core==0.7.1
autogen-ext==0.7.1
certifi==2025.8.3
charset-normalizer==3.4.2
defusedxml==0.7.1
diskcache==5.6.3
distro==1.9.0
docker==7.1.0
Flask==3.0.0
Flask-Cors==4.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
importlib_metadata==8.7.0
jira==3.10.5
jiter==0.10.0
jsonref==1.1.0
oauthlib==3.3.1
openai==1.98.0
opentelemetry-api==1.36.0
packaging==25.0
pillow==11.3.0
protobuf==5.29.5
pydantic==2.11.7
pydantic_core==2.33.2
python-dotenv==1.1.1
python-json-logger==3.3.0
PyYAML==6.0.1
regex==2025.7.34
requests==2.32.4
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
sniffio==1.3.1
termcolor==3.1.0
tiktoken==0.9.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
zipp==3.23.0
